Library

Documentation for SatelliteToolboxAtmosphericModels.jl.

SatelliteToolboxAtmosphericModels.AtmosphericModels.JB2008Output — Type

struct JB2008Output{T<:Number}

Output of the atmospheric model Jacchia-Bowman 2008.

Fields

total_density::T: Total atmospheric density [1 / m³].

temperature::T: Temperature at the selected position [K].

exospheric_temperature::T: Exospheric temperature [K].

N2_number_density::T: Number density of N₂ [1 / m³].

O2_number_density::T: Number density of O₂ [1 / m³].

O_number_density::T: Number density of O [1 / m³].

Ar_number_density::T: Number density of Ar [1 / m³].

He_number_density::T: Number density of He [1 / m³].

H_number_density::T: Number density of H [1 / m³].

SatelliteToolboxAtmosphericModels.AtmosphericModels.JR1971Output — Type

struct JR1971Output{T<:Number}

Output of the atmospheric model Jacchia-Roberts 1971.

Fields

total_density::T: Total atmospheric density [1 / m³].

temperature::T: Temperature at the selected position [K].

exospheric_temperature::T: Exospheric temperature [K].

N2_number_density::T: Number density of N₂ [1 / m³].

O2_number_density::T: Number density of O₂ [1 / m³].

O_number_density::T: Number density of O [1 / m³].

Ar_number_density::T: Number density of Ar [1 / m³].

He_number_density::T: Number density of He [1 / m³].

H_number_density::T: Number density of H [1 / m³].

SatelliteToolboxAtmosphericModels.AtmosphericModels.Nrlmsise00Flags — Type

struct Nrlmsise00Flags

Flags to configure NRLMSISE-00.

Fields

F10_Mean::Bool: F10.7 effect on mean.

time_independent::Bool: Independent of time.

sym_annual::Bool: Symmetrical annual.

sym_semiannual::Bool: Symmetrical semiannual.

asym_annual::Bool: Asymmetrical annual.

asyn_semiannual::Bool: Asymmetrical semiannual.

diurnal::Bool: Diurnal.

semidiurnal::Bool: Semidiurnal.

daily_ap::Bool: Daily AP.

all_ut_long_effects::Bool: All UT/long effects.

longitudinal::Bool: Longitudinal.

ut_mixed_ut_long::Bool: UT and mixed UT/long.

mixed_ap_ut_long::Bool: Mixed AP/UT/long.

terdiurnal::Bool: Terdiurnal.

departures_from_eq::Bool: Departures from diffusive equilibrium.

all_tinf_var::Bool: All TINF variations.

all_tlb_var::Bool: All TLB variations.

all_tn1_var::Bool: All TN1 variations.

all_s_var::Bool: All S variations.

all_tn2_var::Bool: All TN2 variations.

all_nlb_var::Bool: All NLB variations.

all_tn3_var::Bool: All TN3 variations.

turbo_scale_height::Bool: Turbo scale height variations.

SatelliteToolboxAtmosphericModels.AtmosphericModels.Nrlmsise00Output — Type

struct Nrlmsise00Output{T<:Number}

Output structure for NRLMSISE00 model.

Fields

total_density::T: Total mass density [kg / m³].

temperature: Temperature at the selected altitude [K].

exospheric_temperature: Exospheric temperature [K].

N_number_density: Nitrogen number density [1 / m³].

N2_number_density: N₂ number density [1 / m³].

O_number_density: Oxygen number density [1 / m³].

aO_number_density: Anomalous Oxygen number density [1 / m³].

O2_number_density: O₂ number density [1 / m³].

H_number_density: Hydrogen number density [1 / m³].

He_number_density: Helium number density [1 / m³].

Ar_number_density: Argon number density [1 / m³].

Remarks

Anomalous oxygen is defined as hot atomic oxygen or ionized oxygen that can become appreciable at high altitudes (> 500 km) for some ranges of inputs, thereby affection drag on satellites and debris. We group these species under the term Anomalous Oxygen, since their individual variations are not presently separable with the drag data used to define this model component.

SatelliteToolboxAtmosphericModels.AtmosphericModels.exponential — Method

exponential(h::Number) -> Float64

Compute the atmospheric density [kg / m³] at the altitude h [m] above the ellipsoid using the exponential atmospheric model:

                ┌            ┐
                │    h - h₀  │
ρ(h) = ρ₀ . exp │ - ──────── │ ,
                │      H     │
                └            ┘

in which ρ₀, h₀, and H are parameters obtained from tables that depend only on h.

SatelliteToolboxAtmosphericModels.AtmosphericModels.jb2008 — Method

jb2008(instant::DateTime, ϕ_gd::Number, λ::Number, h::Number[, F10::Number, F10ₐ::Number, S10::Number, S10ₐ::Number, M10::Number, M10ₐ::Number, Y10::Number, Y10ₐ::Number, DstΔTc::Number]) -> JB2008Output{Float64}
jb2008(jd::Number, ϕ_gd::Number, λ::Number, h::Number[, F10::Number, F10ₐ::Number, S10::Number, S10ₐ::Number, M10::Number, M10ₐ::Number, Y10::Number, Y10ₐ::Number, DstΔTc::Number]) -> JB2008Output{Float64}

Compute the atmospheric density using the Jacchia-Bowman 2008 (JB2008) model.

This model is a product of the Space Environment Technologies, please, refer to the following website for more information:

http://sol.spacenvironment.net/JB2008/

If we omit all space indices, the system tries to obtain them automatically for the selected day jd or instant. However, the indices must be already initialized using the function SpaceIndices.init().

Arguments

jd::Number: Julian day to compute the model.

instant::DateTime: Instant to compute the model represent using DateTime.

ϕ_gd: Geodetic latitude [rad].

λ: Longitude [rad].

h: Altitude [m].

F10: 10.7-cm solar flux [sfu] obtained 1 day before jd.

F10ₐ: 10.7-cm averaged solar flux using a 81-day window centered on input time obtained 1 day before jd.

S10: EUV index (26-34 nm) scaled to F10.7 obtained 1 day before jd.

S10ₐ: EUV 81-day averaged centered index obtained 1 day before jd.

M10: MG2 index scaled to F10.7 obtained 2 days before jd.

M10ₐ: MG2 81-day averaged centered index obtained 2 day before jd.

Y10: Solar X-ray & Ly-α index scaled to F10.7 obtained 5 days before jd.

Y10ₐ: Solar X-ray & Ly-α 81-day averaged centered index obtained 5 days before jd.

DstΔTc: Temperature variation related to the Dst.

Returns

JB2008Output{Float64}: Structure containing the results obtained from the model.

SatelliteToolboxAtmosphericModels.AtmosphericModels.jr1971 — Method

jr1971(instant::DateTime, ϕ_gd::Number, λ::Number, h::Number[, F10::Number, F10ₐ::Number, Kp::Number]) -> JR1971Output{Float64}
jr1971(jd::Number, ϕ_gd::Number, λ::Number, h::Number[, F10::Number, F10ₐ::Number, Kp::Number]) -> JR1971Output{Float64}

Compute the atmospheric density using the Jacchia-Roberts 1971 model.

If we omit all space indices, the system tries to obtain them automatically for the selected day jd or instant. However, the indices must be already initialized using the function SpaceIndices.init().

Arguments

jd::Number: Julian day to compute the model.

instant::DateTime: Instant to compute the model represent using DateTime.

ϕ_gd::Number: Geodetic latitude [rad].

λ::Number: Longitude [rad].

h::Number: Altitude [m].

F10::Number: 10.7-cm solar flux [sfu].

F10ₐ::Number: 10.7-cm averaged solar flux, 81-day centered on input time [sfu].

Kp::Number: Kp geomagnetic index with a delay of 3 hours.

Returns

JR1971Output{Float64}: Structure containing the results obtained from the model.

SatelliteToolboxAtmosphericModels.AtmosphericModels.nrlmsise00 — Method

nrlmsise00(instant::DateTime, h::Number[, ϕ_gd::Number, λ::Number, F10ₐ::Number, F10::Number, ap::Union{Number, AbstractVector}]; kwargs...) -> Nrlmsise00Output{Float64}
nrlmsise00(jd::Number, h::Number[, ϕ_gd::Number, λ::Number, F10ₐ::Number, F10::Number, ap::Union{Number, AbstractVector}]; kwargs...) -> Nrlmsise00Output{Float64}

Compute the atmospheric density using the NRLMSISE-00 model.

If we omit all space indices, the system tries to obtain them automatically for the selected day jd or instant. However, the indices must be already initialized using the function SpaceIndices.init().

Arguments

instant::DateTime: Instant to compute the model represent using DateTime.

jd::Number: Julian day to compute the model.

h::Number: Altitude [m].

ϕ_gd::Number: Geodetic latitude [rad].

λ::Number: Longitude [rad].

F10ₐ::Number: 10.7-cm averaged solar flux, 90-day centered on input time [sfu].

F10::Number: 10.7-cm solar flux [sfu].

ap::Union{Number, AbstractVector}: Magnetic index, see the section AP for more information.

Keywords

flags::Nrlmsise00Flags: A list of flags to configure the model. For more information, see [Nrlmsise00Flags]@(ref). (Default = Nrlmsise00Flags())

include_anomalous_oxygen::Bool: If true, the anomalous oxygen density will be included in the total density computation. (Default = true)

Returns

Nrlmsise00Output{Float64}: Structure containing the results obtained from the model.

AP

The input variable ap contains the magnetic index. It can be a Number or an AbstractVector.

If ap is a number, it must contain the daily magnetic index.

If ap is an AbstractVector, it must be a vector with 7 dimensions as described below:

Index	Description

1	Daily AP.

2	3 hour AP index for current time.

3	3 hour AP index for 3 hours before current time.

4	3 hour AP index for 6 hours before current time.

5	3 hour AP index for 9 hours before current time.

6	Average of eight 3 hour AP indices from 12 to 33 hours prior to current time.

7	Average of eight 3 hour AP indices from 36 to 57 hours prior to current time.

Extended Help

The densities of O, H, and N are set to 0 below 72.5 km.

The exospheric temperature is set to global average for altitudes below 120 km. The 120 km gradient is left at global average value for altitudes below 72.5 km.

Anomalous oxygen is defined as hot atomic oxygen or ionized oxygen that can become appreciable at high altitudes (> 500 km) for some ranges of inputs, thereby affection drag on satellites and debris. We group these species under the term Anomalous Oxygen, since their individual variations are not presently separable with the drag data used to define this model component.

Notes on Input Variables

F10 and F10ₐ values used to generate the model correspond to the 10.7 cm radio flux at the actual distance of the Earth from the Sun rather than the radio flux at 1 AU. The following site provides both classes of values:

ftp://ftp.ngdc.noaa.gov/STP/SOLAR_DATA/SOLAR_RADIO/FLUX/

F10, F10ₐ, and ap effects are neither large nor well established below 80 km and these parameters should be set to 150, 150, and 4 respectively.

If include_anomalous_oxygen is false, the total_density field in the output is the sum of the mass densities of the species He, O, N₂, O₂, Ar, H, and N, but does not include anomalous oxygen.

If include_anomalous_oxygen is false, the total_density field in the output is the effective total mass density for drag and is the sum of the mass densities of all species in this model including the anomalous oxygen.